Fix for CONTINUE=1 builds for logrotate step

+++ stage3/01-misc-config/00-run.sh
@@ -11,4 +11,6 @@ sed -i 's/127.0.1.1.*'"$CURRENT_HOSTNAME"'/127.0.1.1\t'"$HOSTNAME"'/g' ${ROOTFS_
 sed -i 's/weekly/size 1M/g' ${ROOTFS_DIR}/etc/logrotate.d/rsyslog
 sed -i 's/daily/size 1M/g' ${ROOTFS_DIR}/etc/logrotate.d/rsyslog
 
-mv ${ROOTFS_DIR}/etc/cron.daily/logrotate ${ROOTFS_DIR}/etc/cron.hourly/
+if [ -f ${ROOTFS_DIR}/etc/cron.daily/logrotate ]; then
+	mv ${ROOTFS_DIR}/etc/cron.daily/logrotate ${ROOTFS_DIR}/etc/cron.hourly/
+fi
